--- flutter/wtc/lib/auth/auth.dart
@@ -1,25 +1,0 @@
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-import 'package:wtc/app.dart';
-import 'package:wtc/auth/login.dart';
-
-class AuthPage extends StatelessWidget {
-  const AuthPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: StreamBuilder(
-        stream: FirebaseAuth.instance.authStateChanges(),
-        builder: (context, snapshot){
-          if(snapshot.hasData){
-            return const App();
-          }
-          else{
-            return const LoginPage();
-          }
-        },
-      )
-    );
-  }
-}
